4.06

+++ www/src/App/Route/CartridgeController.php
@@ -16,12 +16,14 @@ class CartridgeController extends AbstractController
     public function new_record(): Response
     {
         if ($this->request->getPostStatus()){
-            $array = $this->request->getPostArrayKey(['id', 'barcode', 'price', 'date', 'model', 'service']);
+            $array = $this->request->getPostArrayKey(['barcode', 'price', 'date', 'model', 'service']);
+            $array["id"] = $this->request->getGet('id');
             $cartridge = new Cartridg();
             $cartridge->createNewRecord($array);
             $array['id'] = $cartridge->returnLastRequest();
             $response = $this->redirect(new RawHtmlView(''),
-                'cartridge/record_apply?check=true&id=' . $array['id'], 302);
+                "cartridge" . DIRECTORY_SEPARATOR .
+                "record_apply?check=true&id=" . $array['id'], 302);
         }else{
             $response = $this->successResponse(new View("Cartridge" . DIRECTORY_SEPARATOR . "new_record"), 200);
         }
@@ -31,13 +33,15 @@ class CartridgeController extends AbstractController
     #[RouteMethod('/edit_cartridge', 'GET')]
     public function edit_cartridge(): Response
     {
-        $array = $this->request->getPostArrayKey(['id', 'barcode', 'price', 'date', 'model', 'service']);
+        $array = $this->request->getPostArrayKey(['barcode', 'price', 'date', 'model', 'service']);
+        $array["id"] = $this->request->getGet('id');
         $cartridge = Cartridg::getById($this->request->getGet('id'));// Возвращает экземпляр класса по Id
         if ($this->request->getPostStatus()){
             $cartridge->setValue($array);
             $cartridge->save();
             $response = $this->redirect(new RawHtmlView(''),
-                "cartridge/record_apply?check=true&id=" . $array['id'], 302);
+                "cartridge" . DIRECTORY_SEPARATOR .
+                "record_apply?check=true&id=" . $array['id'], 302);
         }
         else{
             $response = $this->successResponse(new View("Cartridge" . DIRECTORY_SEPARATOR . "edit_cartridge",
@@ -51,7 +55,7 @@ class CartridgeController extends AbstractController
         $cartridge = Cartridg::getById($this->request->getGet('id'));
         $check = $this->request->getGet('check');
         $response = $this->successResponse(new View("Cartridge" . DIRECTORY_SEPARATOR . "record_apply",
-                ['check' => $check, 'cartridge' => $cartridge]), 200);
+                ['check' => $check, 'cartridge' => $cartridge, 'user' => $this->user]), 200);
         return $response;
     }
 
@@ -60,12 +64,13 @@ class CartridgeController extends AbstractController
     {
 
         $cartridge = Cartridg::getById($this->request->getGet('id'));
-        if ($this->request->getPostStatus()){
+        if ($this->request->getPostStatus() && $cartridge != null){
             $cartridge->delete($this->request->getPostKey('id'));
-            $response = $this->redirect(new RawHtmlView(''), "Cartridge" . DIRECTORY_SEPARATOR .
+            $response = $this->redirect(new RawHtmlView(''), "cartridge" . DIRECTORY_SEPARATOR .
                 "record_apply?check=true", 302);
         }else{
-            $response = $this->successResponse(new View("Cartridge" . DIRECTORY_SEPARATOR . "delete_cartridge",['result' => $cartridge]), 200);
+            $response = $this->successResponse(new View("Cartridge"
+                . DIRECTORY_SEPARATOR . "delete_cartridge",['result' => $cartridge]), 200);
         }
         return $response;
     }
